+++ CHANGELOG.md
@@ -23,6 +23,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
       * Layout of the application board is destroyed for those applications which are fully "DONE" but application activation did not take place 
    * App Release Process
       * validation fix for long description
+      * UseCase update in appstatus api
    * AppOverview
       * Recommended section images not loading fix